tpm2: Advance volatile state buffer pointer towards hash
Advance the volatile state's buffer pointer towards the hash
at the end of the buffer in case we are reading a later
version of a volatile state blob that has extra data that
we did not unmarshal. We will find the expected hash there
that we have to use for comparison.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>